Move some tasks to shared tasks
