+++ datascience/tests/test_pipeline/test_flows/test_distribute_pnos.py
@@ -33,8 +33,6 @@ from src.pipeline.flows.distribute_pnos import (
     create_email,
     create_sms,
     extract_fishing_gear_names,
-    extract_pno_units_ports_and_segments_subscriptions,
-    extract_pno_units_targeting_vessels,
     extract_pnos_to_generate,
     extract_species_names,
     fetch_control_units_contacts,
@@ -1018,53 +1016,6 @@ def pno_pdf_document_to_distribute_verified_assigned(
 
 
 @pytest.fixture
-def pno_units_targeting_vessels():
-    return pd.DataFrame(
-        {
-            "vessel_id": [2, 4, 7],
-            "cfr": ["ABC000542519", None, "___TARGET___"],
-            "control_unit_ids_targeting_vessel": [[4], [1, 2], [4]],
-        }
-    )
-
-
-@pytest.fixture
-def pno_units_ports_and_segments_subscriptions():
-    return pd.DataFrame(
-        {
-            "port_locode": [
-                "FRCQF",
-                "FRDKK",
-                "FRDPE",
-                "FRLEH",
-                "FRLEH",
-                "FRZJZ",
-                "FRZJZ",
-            ],
-            "control_unit_id": [1, 2, 4, 2, 3, 2, 3],
-            "receive_all_pnos_from_port": [
-                False,
-                False,
-                True,
-                False,
-                False,
-                False,
-                False,
-            ],
-            "unit_subscribed_segments": [
-                ["SWW01/02/03"],
-                [],
-                [],
-                [],
-                ["SWW01/02/03", "NWW01"],
-                [],
-                ["SWW01/02/03", "NWW01"],
-            ],
-        }
-    )
-
-
-@pytest.fixture
 def monitorenv_control_units_api_response() -> list:
     return [
         {
@@ -1528,20 +1479,6 @@ def test_extract_pnos_to_generate(reset_test_data, extracted_pnos):
 def test_extract_species_names(reset_test_data, species_names):
     res = extract_species_names.run()
     assert res == species_names
-
-
-def test_extract_pno_units_targeting_vessels(
-    reset_test_data, pno_units_targeting_vessels
-):
-    res = extract_pno_units_targeting_vessels.run()
-    pd.testing.assert_frame_equal(res, pno_units_targeting_vessels)
-
-
-def test_extract_pno_units_ports_and_segments_subscriptions(
-    reset_test_data, pno_units_ports_and_segments_subscriptions
-):
-    res = extract_pno_units_ports_and_segments_subscriptions.run()
-    pd.testing.assert_frame_equal(res, pno_units_ports_and_segments_subscriptions)
 
 
 def test_extract_fishing_gear_names(reset_test_data, fishing_gear_names):
